feat: support for database backend
This changeset adds a database backend interface, with
current support for filesystem (json) and sqlite. If/when
we extend this to others we can bring in sqlalchemy. I
am still working on the validation (manual) logic and
I realize we also need a way for the agent to better
indicate that a step or the workflow is done.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>