feat: support for jinja2 rendering of outputs
When we need to pass forward logic from some previous
step into a prompt (where an LLM is not going to be able
to map arguments) we can use simple jinja2 rendering. This
update also adds the first working examples for a manual and
LLM derived jobspec translation! We are close to being ready
to at least prototype experiments, although I will be afk
for a bit going on adventures for a few days.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>